Replace PlexiBot.java
The old PlexiBot.java was replaced with the code from the new PlexiBotSingleton.java class. The now extra class was deleted. Some refactoring was done to enable Plexi to use the changed PlexiBot class.
